update: refactor FooterControl to use custom hook for dimension handling and remove unused code
